cache-persist, feat(async-storage community, deprecated AsyncStorage react-native)
