+++ packages/cache-persist/src/deprecated/index.ts
@@ -1,0 +1,1 @@
+export {default as storage} from './storage';
