+++ packages/cache-persist/src/storage.native.ts
@@ -1,41 +1,9 @@
-import { AsyncStorage } from 'react-native';
-import { Storage, ItemCache, DataCache } from './CacheTypes';
-
+import AsyncStorage from '@react-native-community/async-storage';
+import { Storage } from './CacheTypes';
+import InternalNativeStorage from './internal/StorageNative'
 
 function NativeStorage(): Storage {
-    return {
-        multiRemove: (keys) => {
-            return AsyncStorage.multiRemove(keys);
-        },
-        multiGet: (keys) => {
-            return AsyncStorage.multiGet(keys).then((data: Array<Array<string>>): DataCache => {
-                const result: DataCache = {};
-                for (var i = 0; i < data.length; i++) {
-                    const itemStorage = data[i];
-                    const key = itemStorage[0];
-                    const value = itemStorage[1];
-                    result[key] = value;
-                }
-                return result;
-            });
-        },
-        getAllKeys: (): Promise<Array<string>> => {
-            return AsyncStorage.getAllKeys();
-        },
-        multiSet: (items: Array<ItemCache<any>>) => {
-            const asyncItems = [];
-            items.forEach(function (item) {
-                asyncItems.push([item.key, item.value])
-            });
-            return AsyncStorage.multiSet(asyncItems);
-        },
-        setItem: (key: string, value: string): Promise<void> => {
-            return AsyncStorage.setItem(key, value)
-        },
-        removeItem: (key: string): Promise<void> => {
-            return AsyncStorage.removeItem(key);
-        },
-    }
+    return InternalNativeStorage(AsyncStorage)
 }
 
 export default NativeStorage;
